Vastly improved and optimised component classes
*Added back in shipID to Component and its inheritors. I realised that I
needed it so a component could access the methods of its owner ship and
its components.
*Reactor is now not relied upon to keep track of where the energy goes.
The reactor doesn't know at all where the energy is going, but there are
now much less lazy ways of setting the power level. It doesn't work in
increments of five, you can set it to whatever you want. I'm not 100%
it'll work yet though.
*Minor move arounds in Ship class.
*Done some intense fleshing out for the Weapon class.