EXTREMELY basic multiplayer.
Basic attacks where clients can invoke damage on another player, however
it's currently broken. I need to overhaul the code again and make sure
that the damage is dealt to the intended players. I also need to make
sure that the target player and any other players who can see the target
player are informed of the change in ship health. I am currently unaware
of how to push data to a specific client without a connection from the
client.