~ Began improving netcode.
*With test data, basic attacks now work and will always be applied to
the client connected on connection 1. I will remove this test data later
and make it applicable to the target.
*Data is now sent to the affected parties when an action is taken upon
them. I have made use of Kryonet's 'getID()' method for this as it
returns the connectionID.
*For now, because I am lazy and like to have less code clogging up my
code, I have made it so the session ID is set to the connection ID. That
way, I don't have to muck around modifying code and I'm able to keep the
ships unique at all times.
*Ships request health and I've added methods for dealing with this. They
will be sent updates on their health whenever they are attacked with
basic attack. I will optimise all this code so it works on any changes.